Enable WifiManager when SSID is configured
- This allows OTA update when a user opts to install firmware
  with configured SSID and password
